Can now render specific tiles and has a cli
The render_specific function allows for rendering tiles of the format
z/x/y. Simply pass a list of files to be generated to it. I've also
added a very basic cli. It should be modified for some mutual exclusion
for bbox/zoom and dirty tiles, but this works for now.